Stutters and size regressions for a GlobalFiler 3500 kit

A dataset containing default parameters and settings for a GlobalFiler 3500 kit.

Usage

gf_configuration()

Format

A list of:

autosomal_markers

Names of autosomal markers in the GlobalFiler kit

repeat_length_by_marker

Named numeric with STR repeat length by locus name

size_regression

See read_size_regression

stutters

List of 4 stutter types, to be used with allele_specific_stutter_model

stutter_model

For convenience, a pre-defined allele_specific_stutter_model

log_normal_settings

Settings corresponding to a log normal model with all stutter types

log_normal_settings_fwbw

Settings corresponding to a log normal model with backward and forward stutter only

gamma_settings

Settings corresponding to a gamma model with all stutter types

gamma_settings_no_stutter

Settings for a gamma model without stutter