Sample drop model(s) with parameters according to priors

Sample drop model(s) with parameters according to priors

Usage

sample_drop_model(
  number_of_contributors,
  sampling_parameters,
  drop_in_rate = 0,
  model_settings
)

Arguments

number_of_contributors

Integer

sampling_parameters

List. Needs to contain:

min_dropout_probability. Numeric of length one.

max_dropout_probability Numeric of length one.

drop_in_rate

Numeric vector of length one. Expected number of drop-ins per locus. Default is 0.

model_settings

List. See drop_model.

Value

When length(number_of_contributors)==1, a single drop_model of class pg_model. Otherwise, a list of these.

Details

In simulation studies involving many mixed DNA profiles, one often needs to generate various samples with different model parameters. This function samples a drop model with parameters according to prior distributions. The dropout probability for each contributor is sampled uniformly between min_dropout_probability. and max_dropout_probability.

See also

Examples

gf <- gf_configuration()

sampling_parameters <- list(min_dropout_probability. = 0., max_dropout_probability. = 0.5)

model <- sample_drop_model(number_of_contributors = 1,
                           sampling_parameters = sampling_parameters,
                           model_settings = list(locus_names = gf$autosomal_markers,
                           size_regression = gf$size_regression))